순환참조 에러 구조수정 작업 완료

+++ project/back-end/src/main/java/com/backend/project/dto/SearchDTO.java
@@ -10,13 +10,14 @@ import org.hibernate.annotations.Type;
 @Setter
 @NoArgsConstructor
 public class SearchDTO {
+
     @Type(type = "text")
     private String keyword;
 
     @Type(type = "text")
     private String classification;
 
-    private Long offsetCount;
+    private int offsetCount;
 
     private Account account;
